creacion de metodo para crear dates automatizados

+++ service-dateEntity/src/main/java/com/example/service_dateEntity/model/DailyAvailability.java
@@ -28,7 +28,7 @@ public class DailyAvailability {
 
     private Integer duration;
 
-    private List<Integer> range;
+    private List<Double> range;
 
     private Integer rest;
